Remove &egui::Context params
